requirements.txt
gTTS
python-telegram-bot
google-generativeai
python-dotenv
